+++ c/src/tests/psxtests/psx06/task2.c
@@ -38,7 +38,7 @@ void *Task_2(
  
   key_data = pthread_getspecific( Key_id );
   printf( "Task_2: Got the key value of %ld\n",
-          (rtems_unsigned32 *)key_data - Data_array );
+          (unsigned long) ((rtems_unsigned32 *)key_data - Data_array) );
   if ( status )
     printf( "status = %d\n", status );
   assert( !status );
