+++ testsuites/psxtests/psx06/task.c
@@ -37,7 +37,7 @@ void *Task_1(
  
   key_data = pthread_getspecific( Key_id );
   printf( "Task_1: Got the key value of %ld\n",
-          (rtems_unsigned32 *)key_data - Data_array );
+          (unsigned long) ((rtems_unsigned32 *)key_data - Data_array) );
   if ( status )
     printf( "status = %d\n", status );
   assert( !status );
